chatbot/api_selector/requirements.txt
chromadb-client
pandas
joblib==1.3.2
langchain==0.1.13
langchain-openai==0.1.1
openapi-core==0.19.0
